fix: changes new commands

+++ commands/say.js
@@ -1,7 +1,11 @@
 const Discord = require("discord.js");
 
 module.exports = { 
-  run: async (client, message, args) => {
+  run: async ({client, message, args}) => {
+    if(message.author.username != `douglascastrorj`) {
+      message.channel.send('Ta achando que manda em mim seu otario?');
+      return;
+    }
     console.log('say message', message);
     const sayMessage = args.join(" ");
     message.delete().catch(O_o => {})
